Put intent functions in own object
